+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.5.0] - 2022-03-30
 
 ### Added
 - Add Maven BOM [#125](https://github.com/nbbrd/java-service-util/issues/125)
@@ -74,7 +74,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/java-service-util/compare/v1.4.0...HEAD
+[Unreleased]: https://github.com/nbbrd/java-service-util/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/nbbrd/java-service-util/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/nbbrd/java-service-util/compare/v1.3.2...v1.4.0
 [1.3.2]: https://github.com/nbbrd/java-service-util/compare/v1.3.0...v1.3.2
 [1.3.0]: https://github.com/nbbrd/java-service-util/compare/v1.2.1...v1.3.0